app_kivy/requirements.txt
kivy
kivymd
python-dotenv
flask
buildozer
pyjnius
cython